Modifications on component percentStateChangeCard
